+++ CactusEngine/Graphics/Resources/Plane.h
@@ -6,7 +6,7 @@ namespace Engine
 	class Plane : public Mesh
 	{
 	public:
-		Plane(uint32_t dimLength, uint32_t dimWidth); // The dimension is grid count (1, 2, 3, ...)
+		Plane(uint64_t dimLength, uint64_t dimWidth); // The dimension is grid count (1, 2, 3, ...)
 		~Plane() = default;
 	};
 }
